requirements.txt
PyQt6
requests
pydantic
fastapi
python-jose[cryptography]
passlib
uvicorn